[testBug] move table qualifier test to a better place
