+++ jpa/eclipselink.jpa.test/src/org/eclipse/persistence/testing/tests/jpa/advanced/AdvancedJunitTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 1998, 2014 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2019 Oracle and/or its affiliates. All rights reserved.
  * This program and the accompanying materials are made available under the 
  * terms of the Eclipse Public License v1.0 and Eclipse Distribution License v. 1.0 
  * which accompanies this distribution. 
@@ -46,7 +46,6 @@ import org.eclipse.persistence.testing.models.jpa.advanced.VegetablePK;
 import org.eclipse.persistence.testing.models.jpa.advanced.WorldRank;
 import org.eclipse.persistence.testing.models.jpa.advanced.PartnerLink;
 import org.eclipse.persistence.testing.models.jpa.advanced.LargeProject;
-import org.eclipse.persistence.testing.models.jpa.advanced.entities.EntityWithSchema;
 import org.eclipse.persistence.testing.models.jpa.advanced.entities.SimpleEntity;
 import org.eclipse.persistence.testing.models.jpa.advanced.entities.SimpleNature;
 import org.eclipse.persistence.testing.models.jpa.advanced.entities.SimpleLanguage;
@@ -75,7 +74,6 @@ public class AdvancedJunitTest extends JUnitTestCase {
         suite.addTest(new AdvancedJunitTest("testElementCollectionClear"));
         suite.addTest(new AdvancedJunitTest("testElementCollectionEntityMapKeyRemove"));
         suite.addTest(new AdvancedJunitTest("testSwitchBatchDuringSessionEvent"));
-        suite.addTest(new AdvancedJunitTest("testTableQualifierOnSequencingTableAndEntity"));
         
         return suite;
     }
@@ -529,35 +527,5 @@ public class AdvancedJunitTest extends JUnitTestCase {
             clearCache();
             closeEntityManager(em);
         }
-    }
-    
-    /*
-     * Test for Bug 445466
-     * When a table qualifier is configured on an Entity/TableSequence combination, 
-     * the SQL is not printed correctly. Programmatically test that an Entity's TableSequence 
-     * has the expected table name & qualifier information configured. 
-     */
-    public void testTableQualifierOnSequencingTableAndEntity() {
-        // Access serverSession directly, do not create an EM as the Entity 'EntityWithSchema' 
-        // has a hard-coded schema for programmatic testing only
-        ClassDescriptor descriptor = getServerSession(getPersistenceUnitName()).getDescriptor(EntityWithSchema.class);
-        TableSequence tableSequence = (TableSequence) descriptor.getSequence();
-        DatabaseTable table = tableSequence.getTable();
-        
-        String tableName = table.getName(); // returns fully qualified table name, by design
-        String tableQualifier = table.getTableQualifier();
-        
-        String tsQualifier = tableSequence.getQualifier();
-        String tsTableName = tableSequence.getTableName();
-        String tsQualifiedTableName = tableSequence.getQualifiedTableName();
-        
-        assertNotNull("Table name should be non-null", tableName);
-        assertTrue("Table name should not be empty", tableName.length() > 0);
-        assertNotNull("Table qualifier should be non-null", tableQualifier);
-        assertTrue("Table qualifier should not be empty", tableQualifier.length() > 0);
-        
-        assertEquals("Table Sequence : table name (with qualifier) should be equal", (tableQualifier + "." + tableName), tsTableName);
-        assertEquals("Table Sequence : table qualifier should be equal", tableQualifier, tsQualifier);
-        assertEquals("Table Sequence : qualified table name should be equal", (tableQualifier + "." + tableName), tsQualifiedTableName);
     }
 }
